RavenDB-22835 : address PR comments - move cancellation token from Task.Run into CountersRepairTask, move FixCountersForDocuments from CountersStorage to CountersRepairTask
